Makefile works on Windows using MinGW

+++ Makefile
@@ -44,6 +44,8 @@ default:
 #################################################################
 # Compiler/Linker options for all build targets
 #################################################################
+
+# Use MinGW on windows
 CC      = gcc
 BIN     = bin
 OUT     = obj
@@ -87,10 +89,18 @@ HEADERS_X11 = \
 # Clean either temporary files or all of it
 #################################################################
 clean:
+ifeq ($(OS),Windows_NT)
+	if exist $(OUT)\nul rmdir /s /q $(OUT)
+else
 	rm -rf $(OUT)
+endif
 
 clean-all: clean
+ifeq ($(OS),Windows_NT)
+	if exist $(BIN)\nul rmdir /s /q $(BIN)
+else
 	rm -rf $(BIN)
+endif
 
 
 #################################################################
@@ -188,7 +198,7 @@ $(OUT)/osx-x11/groufix.o: $(SRC)/groufix.c $(HEADERS_X11)
 #################################################################
 # Windows builds
 #################################################################
-OBJS_OSX_X11 = \
+OBJS_WIN32 = \
  $(OUT)/win32/groufix/platform/win32_context.o \
  $(OUT)/win32/groufix/platform/win32_init.o \
  $(OUT)/win32/groufix/platform/win32_screen.o \
@@ -197,13 +207,15 @@ OBJS_OSX_X11 = \
  $(OUT)/win32/groufix/math.o \
  $(OUT)/win32/groufix.o
 
-win32: before-win32
+win32: before-win32 $(OBJS_WIN32)
+	$(CC) -shared $(OBJS_WIN32) -o $(BIN)/win32/libGroufix.dll -luser32 -lopengl32
 
-win32-simple: win32
+win32-simple: examples/simple/main.c win32
+	$(CC) $(CFLAGS) $< -o $(BIN)/win32/simple -L$(BIN)/win32/ -lGroufix
 
 before-win32:
-	mkdir -p $(BIN)/win32
-	mkdir -p $(OUT)/win32/groufix/platform
+	if not exist $(BIN)\nul mkdir $(BIN)\win32
+	if not exist $(OUT)\nul mkdir $(OUT)\win32\groufix\platform
 
 
 # All the object files
